# ai-server/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Actual Intelligence API-DOC",
    "version": "1.0.0",
    "description": "API-Dokumentation für das Chatroom Backend.\n\n## 🔌 WebSocket Verbindung\n\nDie Echtzeit-Kommunikation läuft über WebSockets unter `ws://localhost:3000`.\n\n### Unterstützte Events (Client an Server)\nDer Server erwartet JSON-Payloads im folgenden Format:\n\n```json\n{\n  \"event\": \"join\" | \"startTyping\" | \"stopTyping\" | \"message\",\n  \"chat_id\": 123,\n  \"username\": \"Max\",\n  \"message\": \"Hallo!\" // (Nur bei event: message)\n}\n```\n\n### Antworten (Server an Client)\nDer Server sendet Nachrichten im Format:\n```json\n{\n  \"event\": \"message\"| \"stopTyping\" | \"startTyping\" | \"error\",\n  \"data\": { ... }\n}\n```"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Lokaler Entwicklungsserver"
    }
  ],
...
  "paths": {
...
    "/prompt": {
      "get": {
        "tags": ["Prompts"],
        "summary": "Alle Prompts abrufen",
        "responses": {
          "200": {
            "description": "Liste aller Prompts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Prompt"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": ["Prompts"],
        "summary": "Einen neuen Prompt erstellen",
        "security": [
          {
            "accessToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewPrompt"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Prompt erfolgreich erstellt"
          },
          "400": {
            "description": "Missing or invalid fields | Failed to create prompt or Category does not exist"
          },
          "401": {
            "description": "Authentication required | Invalid or expired token"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/prompt/{id}": {
      "get": {
        "tags": ["Prompts"],
        "summary": "Prompt über ID abrufen",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Der gewünschte Prompt",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Prompt"
                }
              }
            }
          },
          "400": {
            "description": "Invalid prompt id"
          },
          "404": {
            "description": "Prompt with ID ${id} not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "accessToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "Wird für die meisten Routen benötigt. Wird beim Login gesetzt."
      },
      "refreshToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken",
        "description": "Wird nur für die /refresh Route benötigt."
      }
    },
    "schemas": {
...
      "Prompt": {
        "type": "object",
        "properties": {
          "prompt_id": {
            "type": "integer"
          },
          "category_id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "time_stamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "NewPrompt": {
        "type": "object",
        "required": ["category_id", "title", "description"],
        "properties": {
          "category_id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
    }
  }
}